pyproject.toml
[project]
name = "how-fast"
version = "0.1.0"
description = "Inference benchmarking library for NanoServe"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28",
    "pyyaml>=6",
    "pydantic>=2",
    "numpy>=1.26",
    "pandas>=2",
]

[project.optional-dependencies]
deploy = ["modal>=0.73"]
notebook = ["matplotlib>=3.8", "plotly>=5.18", "jupyter>=1.0"]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[project.scripts]
how-fast = "how_fast.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/how_fast"]
